pyproject.toml
[project]
name = "main-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "ipykernel>=7.1.0",
    "kagglehub[pandas-datasets]>=0.3.13",
    "mlflow>=3.8.1",
    "pandas>=2.3.3",
    "pip>=25.3",
    "pyarrow>=22.0.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "xgboost>=3.1.3",
]
